move v2TablePrivileges method to v2Commands
